backend/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-environ==0.4.5
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.4
djangorestframework-jwt
Markdown==3.3.4
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1